refactor(tools): extract shared doDSQuery helper for /api/ds/query
The VictoriaMetrics and Cloud Monitoring backends had a verbatim copy
of the same /api/ds/query helper. Lift it to a free function in
prom_backend.go so future fixes apply to both.
